+++ packages/core/src/StateMachine.ts
@@ -6,6 +6,7 @@ import { StateNode } from './StateNode.ts';
 import { interpret } from './interpreter.ts';
 import {
   getConfiguration,
+  getStateNodeByPath,
   getInitialConfiguration,
   getStateNodes,
   isInFinalState,
@@ -371,9 +372,11 @@ export class StateMachine<
   }
 
   public getStateNodeById(stateId: string): StateNode<TContext, TEvent> {
-    const resolvedStateId = isStateId(stateId)
-      ? stateId.slice(STATE_IDENTIFIER.length)
-      : stateId;
+    const fullPath = stateId.split(this.delimiter);
+    const relativePath = fullPath.slice(1);
+    const resolvedStateId = isStateId(fullPath[0])
+      ? fullPath[0].slice(STATE_IDENTIFIER.length)
+      : fullPath[0];
 
     const stateNode = this.idMap.get(resolvedStateId);
     if (!stateNode) {
@@ -381,7 +384,7 @@ export class StateMachine<
         `Child state node '#${resolvedStateId}' does not exist on machine '${this.id}'`
       );
     }
-    return stateNode;
+    return getStateNodeByPath(stateNode, relativePath);
   }
 
   public get definition(): StateMachineDefinition<TContext, TEvent> {