+++ packages/core/test/parallel.test.ts
@@ -759,7 +759,7 @@ describe('parallel states', () => {
         }
       },
       on: {
-        'to-A': 'A'
+        'to-A': '.A'
       }
     });